wip: login UI, switched to oauth
